+++ tests/kernel/errno/prj.conf
@@ -1,1 +1,1 @@
-# nothing yet
+CONFIG_ZTEST=y
