+++ tests/kernel/errno/prj_newlib.conf
@@ -1,1 +1,2 @@
 CONFIG_NEWLIB_LIBC=y
+CONFIG_ZTEST=y
